+++ src/app/chat/chat.scss
@@ -1,0 +1,24 @@
+@import '../variables.scss';
+
+// .space-container {
+//   display: grid;
+//   grid-template-areas:
+//     "sidebar chatwindow"
+//     "sidebar chatwindow";
+//   height: #{$view-container-height-no-toolbar};
+// }
+
+.space-container {
+  display: flex;
+  flex-direction: column;
+  height: #{$view-container-height-no-toolbar};
+  overflow: hidden;
+}
+
+// .sidebar-container {
+//   grid-area: sidebar;
+// }
+
+// .window-container {
+//   grid-area: chatwindow;
+// }
